+++ app-src/lib/browser/FullScreen/FullScreen.js
