+++ app-src/lib/browser/FullScreen/index.js
