+++ app-src/lib/browser/LocalStorage/index.js
