+++ app-src/lib/index.js
@@ -1,0 +1,4 @@
+import browser from './browser';
+import radioService, { initialiseRadio } from './radioService';
+
+export { browser, radioService, initialiseRadio };
