+++ app-src/lib/radioService/index.js
@@ -1,0 +1,1 @@
+export { radioService as default, initialiseRadio } from './radioService';
