--- app-src/services/index.js
@@ -1,4 +1,0 @@
-import browser from './browser';
-import radio, { initialiseRadio } from './radio';
-
-export { browser, radio, initialiseRadio };
