--- app-src/services/radio/index.js
@@ -1,1 +1,0 @@
-export { radio as default, initialiseRadio } from './radio';
